resourcePath correction & alignment
There are two different places from which you can access the resource
path: within the layout as a synthetic property of the frontmatter,
and when the frontmatter is imported into a different file, where
__resourcePath is also available as a synthetic property.

Previously these two paths didn't align perfectly - when accessing
it from a layout, it would be the path relative to 'pages' with no
leading slash, but when importing it would have a leading slash.
This commit makes them match exactly.